feat: implement post authorization with policies

+++ app/Http/Controllers/PostController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Gate;
+
 use App\Models\Post;
 use App\Http\Resources\Post as PostResource;
 use App\Http\Requests\PostRequest;
@@ -14,6 +16,8 @@ class PostController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Post::class);
+
         $posts = Post::orderBy('created_at', 'desc')->paginate(10);
 
         return new PostCollectionResource($posts);
@@ -24,6 +28,8 @@ class PostController extends Controller
      */
     public function store(PostRequest $request)
     {
+        Gate::authorize('create', Post::class);
+
         $post = $request
             ->user()
             ->posts()
@@ -37,6 +43,8 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        Gate::authorize('view', $post);
+
         return new PostResource($post);
     }
 
@@ -45,6 +53,8 @@ class PostController extends Controller
      */
     public function update(PostRequest $request, Post $post)
     {
+        Gate::authorize('update', $post);
+
         $post->update($request->all());
 
         return new PostResource($post);
@@ -55,6 +65,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        Gate::authorize('delete', $post);
+
         $post->delete();
 
         return response()->noContent();
